+++ src/main/java/site/travellaboratory/be/article/application/service/ArticleLikeService.java
@@ -1,4 +1,4 @@
-package site.travellaboratory.be.article.application;
+package site.travellaboratory.be.article.application.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
