+++ src/main/java/site/travellaboratory/be/article/application/service/_schedule/ArticleScheduleWriterService.java
@@ -1,4 +1,4 @@
-package site.travellaboratory.be.article.application._schedule;
+package site.travellaboratory.be.article.application.service._schedule;
 
 import java.util.List;
 import java.util.Map;
