+++ src/main/java/site/travellaboratory/be/article/presentation/controller/ArticleLikeController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import site.travellaboratory.be.article.application.ArticleLikeService;
+import site.travellaboratory.be.article.application.service.ArticleLikeService;
 import site.travellaboratory.be.common.annotation.UserId;
 import site.travellaboratory.be.article.presentation.response.like.BookmarkSaveResponse;
